requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets

nemo_toolkit[asr]
torch>=2.0.0
torchaudio
silero
webrtcvad

pydub
soundfile
numpy==1.26.4

selenium
scikit-learn
requests